support MAIL command with a null
MAIL command with a null is required in rfc5321 as shown below.

https://tools.ietf.org/html/rfc5321#section-3.6.3
```
One way to prevent loops in error reporting is to specify a null reverse-path
in the MAIL command of a notification message.
When such a message is transmitted, the reverse-path MUST be set to null
(see   Section 4.5.5 for additional discussion).
A MAIL command with a null reverse-path appears as follows:

  MAIL FROM:<>
```

In fact, Postfix bounce emails have MAIL command set to "FROM:<>".
But, MailHog validation judges "FROM:<>" as an error.
Therefore, Postfix bounce email cannot be received by MailHog.

This Pull Request allows you to receive Postfix bounce email by support MAIL command with a null.